% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{shinify}
\alias{shinify}
\title{Shiny Server}
\usage{
shinify(model, modeltype = "", title = "")
}
\arguments{
\item{modeltype}{type of your model. Currently for "log_reg", "lin_reg","df_rpart", "dt_party" and "rf"}

\item{model}{Your R model}
}
\description{
This function creates a shiny server for your model
}
\examples{
shinify(model)
shinify(model, "log_reg")
shinify(model, "log_reg", "your awesome title")
}
\keyword{shiny}
